Bullets collisions work and do damage.
You can kill the test enemy (orange circle thing) now if you hit it 5 times.

The shooting needs to be modified for all 4 players since right now they all shoot towards player 1's target.

Right now the player spawns the bullet, but really it will need to be the weapon spawning the bullet. Shouldn't be a bad change, but it means there's no point in fixing the multiplayer shooting bug until the design is finalized.