Fixed some bugs in menus, updated NumOfPlayers script to test 4plyrs
Updated NumOfPlayers script to allow testing up to four players.  This is the script that's attached to the "SceneControl" object in Level01, if you enable the SceneControl object, you can set the number of players and set the class of each player.  Each player shares the same keyboard controls, but it's good for testing multiplayer.  I also fixed some bugs within the menus.